+++ .eslintrc.js
@@ -13,5 +13,7 @@ module.exports = {
   },
   rules: {
     "node/no-unsupported-features/es-syntax": ["error", { ignores: ["modules"] }],
+    "mocha/no-exclusive-tests": "error",
+    "@typescript-eslint/no-var-requires": 0,
   },
 };
