+++ test/HubPool.LiquidityProvisionFees.ts
@@ -1,6 +1,6 @@
 import { expect, ethers, Contract, SignerWithAddress, seedWallet, toWei } from "./utils";
 import * as consts from "./constants";
-import { hubPoolFixture, enableTokensForLP } from "./HubPool.Fixture";
+import { hubPoolFixture, enableTokensForLP } from "./fixtures/HubPool.Fixture";
 import { constructSingleChainTree } from "./MerkleLib.utils";
 
 let hubPool: Contract, weth: Contract, timer: Contract;
