+++ test/SpokePool.ExecuteRootBundle.ts
@@ -1,12 +1,7 @@
-import { expect } from "chai";
-import { Contract } from "ethers";
-import { ethers } from "hardhat";
-
-import { SignerWithAddress, seedContract, toBN, toWei } from "./utils";
+import { SignerWithAddress, seedContract, toBN, expect, Contract, ethers } from "./utils";
 import * as consts from "./constants";
-import { spokePoolFixture } from "./SpokePool.Fixture";
+import { spokePoolFixture } from "./fixtures/SpokePool.Fixture";
 import { buildRelayerRefundTree, buildRelayerRefundLeafs } from "./MerkleLib.utils";
-import { toBNWei } from "@uma/common";
 
 let spokePool: Contract, destErc20: Contract, weth: Contract;
 let dataWorker: SignerWithAddress, relayer: SignerWithAddress, rando: SignerWithAddress;
@@ -78,27 +73,6 @@ describe("SpokePool Root Bundle Execution", function () {
     expect(tokensBridgedEvents.length).to.equal(1);
   });
 
-  // TODO: Move to Optimism_SpokePool test.
-  // it("Execute relayer root wraps any ETH owned by contract", async function () {
-  //   const { leafs, tree } = await constructSimpleTree(destErc20, destinationChainId);
-
-  //   // Store new tree.
-  //   await spokePool.connect(dataWorker).relayRootBundle(
-  //     tree.getHexRoot(), // relayer refund root. Generated from the merkle tree constructed before.
-  //     consts.mockSlowRelayRoot
-  //   );
-
-  //   const amountOfEthToWrap = toWei("1");
-  //   await rando.sendTransaction({
-  //     to: spokePool.address,
-  //     value: amountOfEthToWrap,
-  //   });
-
-  //   // Pool should have wrapped all ETH
-  //   await expect(() =>
-  //     spokePool.connect(dataWorker).executeRelayerRefundRoot(0, leafs[0], tree.getHexProof(leafs[0]))
-  //   ).to.changeEtherBalance(spokePool, amountOfEthToWrap.mul(-1));
-  // });
   it("Execution rejects invalid leaf, tree, proof combinations", async function () {
     const { leafs, tree } = await constructSimpleTree(destErc20, destinationChainId);
     await spokePool.connect(dataWorker).relayRootBundle(
