+++ test/chain-adapters/Arbitrum_Adapter.ts
@@ -1,7 +1,7 @@
 import * as consts from "../constants";
 import { ethers, expect, Contract, FakeContract, SignerWithAddress, createFake, toWei, hre } from "../utils";
 import { getContractFactory, seedWallet, randomAddress } from "../utils";
-import { hubPoolFixture, enableTokensForLP } from "../HubPool.Fixture";
+import { hubPoolFixture, enableTokensForLP } from "../fixtures/HubPool.Fixture";
 import { constructSingleChainTree } from "../MerkleLib.utils";
 
 let hubPool: Contract,
