+++ test/chain-adapters/Optimism_Adapter.ts
@@ -1,14 +1,7 @@
-import {
-  sampleL2Gas,
-  amountToLp,
-  mockTreeRoot,
-  refundProposalLiveness,
-  bondAmount,
-  mockSlowRelayRoot,
-} from "./../constants";
+import { sampleL2Gas, amountToLp, mockTreeRoot, refundProposalLiveness, bondAmount } from "./../constants";
 import { ethers, expect, Contract, FakeContract, SignerWithAddress, createFake, hre } from "../utils";
 import { getContractFactory, seedWallet, randomAddress } from "../utils";
-import { hubPoolFixture, enableTokensForLP } from "../HubPool.Fixture";
+import { hubPoolFixture, enableTokensForLP } from "../fixtures/HubPool.Fixture";
 import { constructSingleChainTree } from "../MerkleLib.utils";
 
 let hubPool: Contract,
@@ -73,7 +66,7 @@ describe("Optimism Chain Adapter", function () {
     // Create an action that will send an L1->L2 tokens transfer and bundle. For this, create a relayer repayment bundle
     // and check that at it's finalization the L2 bridge contracts are called as expected.
     const { leafs, tree, tokensSendToL2 } = await constructSingleChainTree(dai.address, 1, optimismChainId);
-    await hubPool.connect(dataWorker).proposeRootBundle([3117], 1, tree.getHexRoot(), mockTreeRoot, mockSlowRelayRoot);
+    await hubPool.connect(dataWorker).proposeRootBundle([3117], 1, tree.getHexRoot(), mockTreeRoot, mockTreeRoot);
     await timer.setCurrentTime(Number(await timer.getCurrentTime()) + refundProposalLiveness + 1);
     await hubPool.connect(dataWorker).executeRootBundle(leafs[0], tree.getHexProof(leafs[0]));
 
@@ -84,7 +77,7 @@ describe("Optimism Chain Adapter", function () {
     expect(l1StandardBridge.depositERC20To).to.have.been.calledWith(...expectedErc20L1ToL2BridgeParams);
     const expectedL2ToL1FunctionCallParams = [
       mockSpoke.address,
-      mockSpoke.interface.encodeFunctionData("relayRootBundle", [mockTreeRoot, mockSlowRelayRoot]),
+      mockSpoke.interface.encodeFunctionData("relayRootBundle", [mockTreeRoot, mockTreeRoot]),
       sampleL2Gas,
     ];
     expect(l1CrossDomainMessenger.sendMessage).to.have.been.calledWith(...expectedL2ToL1FunctionCallParams);
@@ -92,7 +85,7 @@ describe("Optimism Chain Adapter", function () {
   it("Correctly unwraps WETH and bridges ETH", async function () {
     // Cant bridge WETH on optimism. Rather, unwrap WETH to ETH then bridge it. Validate the adapter does this.
     const { leafs, tree } = await constructSingleChainTree(weth.address, 1, optimismChainId);
-    await hubPool.connect(dataWorker).proposeRootBundle([3117], 1, tree.getHexRoot(), mockTreeRoot, mockSlowRelayRoot);
+    await hubPool.connect(dataWorker).proposeRootBundle([3117], 1, tree.getHexRoot(), mockTreeRoot, mockTreeRoot);
     await timer.setCurrentTime(Number(await timer.getCurrentTime()) + refundProposalLiveness + 1);
     await hubPool.connect(dataWorker).executeRootBundle(leafs[0], tree.getHexProof(leafs[0]));
 
@@ -102,7 +95,7 @@ describe("Optimism Chain Adapter", function () {
     expect(l1StandardBridge.depositETHTo).to.have.been.calledWith(mockSpoke.address, sampleL2Gas, "0x");
     const expectedL2ToL1FunctionCallParams = [
       mockSpoke.address,
-      mockSpoke.interface.encodeFunctionData("relayRootBundle", [mockTreeRoot, mockSlowRelayRoot]),
+      mockSpoke.interface.encodeFunctionData("relayRootBundle", [mockTreeRoot, mockTreeRoot]),
       sampleL2Gas,
     ];
     expect(l1CrossDomainMessenger.sendMessage).to.have.been.calledWith(...expectedL2ToL1FunctionCallParams);
