+++ test/chain-adapters/Polygon_Adapter.ts
@@ -1,24 +1,7 @@
-import {
-  sampleL2Gas,
-  amountToLp,
-  mockTreeRoot,
-  refundProposalLiveness,
-  bondAmount,
-  mockSlowRelayRoot,
-} from "./../constants";
-import {
-  ethers,
-  expect,
-  Contract,
-  FakeContract,
-  SignerWithAddress,
-  createFake,
-  getContractFactory,
-  seedWallet,
-  randomAddress,
-  hre,
-} from "../utils";
-import { hubPoolFixture, enableTokensForLP } from "../HubPool.Fixture";
+import { amountToLp, mockTreeRoot, refundProposalLiveness, bondAmount, mockSlowRelayRoot } from "./../constants";
+import { ethers, expect, Contract, FakeContract, SignerWithAddress } from "../utils";
+import { createFake, getContractFactory, seedWallet, randomAddress, hre } from "../utils";
+import { hubPoolFixture, enableTokensForLP } from "../fixtures/HubPool.Fixture";
 import { constructSingleChainTree } from "../MerkleLib.utils";
 
 let hubPool: Contract,
