+++ test/chain-specific-spokepools/Polygon_SpokePool.ts
@@ -1,26 +1,13 @@
 import { TokenRolesEnum, ZERO_ADDRESS } from "@uma/common";
 import { mockTreeRoot, amountToReturn, amountHeldByPool } from "../constants";
 import { ethers, expect, Contract, SignerWithAddress, getContractFactory, seedContract, toWei } from "../utils";
-import { hubPoolFixture } from "../HubPool.Fixture";
-import { buildRelayerRefundLeafs, buildRelayerRefundTree } from "../MerkleLib.utils";
+import { hubPoolFixture } from "../fixtures/HubPool.Fixture";
+import { constructSingleRelayerRefundTree } from "../MerkleLib.utils";
 
 let hubPool: Contract, polygonSpokePool: Contract, timer: Contract, dai: Contract, weth: Contract;
 
 let owner: SignerWithAddress, relayer: SignerWithAddress, rando: SignerWithAddress, fxChild: SignerWithAddress;
 
-async function constructSimpleTree(l2Token: Contract | string, destinationChainId: number) {
-  const leafs = buildRelayerRefundLeafs(
-    [destinationChainId], // Destination chain ID.
-    [amountToReturn], // amountToReturn.
-    [l2Token as string], // l2Token.
-    [[]], // refundAddresses.
-    [[]] // refundAmounts.
-  );
-
-  const tree = await buildRelayerRefundTree(leafs);
-
-  return { leafs, tree };
-}
 describe("Polygon Spoke Pool", function () {
   beforeEach(async function () {
     [owner, relayer, fxChild, rando] = await ethers.getSigners();
@@ -112,7 +99,10 @@ describe("Polygon Spoke Pool", function () {
   });
 
   it("Bridge tokens to hub pool correctly sends tokens through the PolygonTokenBridger", async function () {
-    const { leafs, tree } = await constructSimpleTree(dai.address, await polygonSpokePool.callStatic.chainId());
+    const { leafs, tree } = await constructSingleRelayerRefundTree(
+      dai.address,
+      await polygonSpokePool.callStatic.chainId()
+    );
     const relayRootBundleData = polygonSpokePool.interface.encodeFunctionData("relayRootBundle", [
       tree.getHexRoot(),
       mockTreeRoot,
